Extract generator's tests from utility's tests
